refactor(streaming): isolate Pi drain wait arbitration
